[ty] Move `UnionType` and `IntersectionType` to a new `types::set_theoretic` submodule
